+++ packages/sdk-scripts/src/cadence/nft/rarible-nft-v2.ts
@@ -1,4 +1,4 @@
-export const RaribleNftV2 = {
+export const RaribleNFTv2 = {
 	mint: `
 import NonFungibleToken from address
 import SoftCollection from address