+++ packages/sdk-scripts/src/index.ts
@@ -1,6 +1,7 @@
 export { commonNft } from "./cadence/nft/common"
 export { RaribleNFT } from "./cadence/nft/rarible-nft"
-export { RaribleNftV2 } from "./cadence/nft/rarible-nft-v2"
+export { RaribleNFTv2 } from "./cadence/nft/rarible-nft-v2"
+export { SoftCollection } from "./cadence/nft/soft-collection"
 
 export { Storefront } from "./cadence/storefront/storefront"
 export { StorefrontCommon } from "./cadence/storefront/storefront-common"