+++ packages/sdk/src/common/collection/index.test.ts
@@ -1,4 +1,5 @@
 import { toFlowContractAddress } from "../flow-address"
+import type { NonFungibleContracts } from "../../types"
 import { getCollectionData, isFlowCollection } from "."
 
 describe("collection", () => {
@@ -17,4 +18,29 @@ describe("collection", () => {
 		const assertion = () => getCollectionData(toFlowContractAddress("A.0x01658d9b94068f3c.Top1Shot"))
 		expect(assertion).toThrow(Error)
 	})
+	test("should check collection data ", () => {
+		collections.forEach(c => {
+			const { name } = getCollectionData(toFlowContractAddress(`A.0x01658d9b94068f3c.${c}`))
+			expect(name).toEqual(c)
+		})
+	})
+
 })
+
+const collections: NonFungibleContracts = [
+	"RaribleNFT",
+	"Evolution",
+	"MotoGPCard",
+	"TopShot",
+	"MugenNFT",
+	"CNN_NFT",
+	"MatrixWorldFlowFestNFT",
+	"MatrixWorldVoucher",
+	"DisruptArt",
+	"Art",
+	"StarlyCard",
+	"OneFootballCollectible",
+	"ChainmonstersRewards",
+	"RaribleNFTv2",
+	"SoftCollection",
+]