+++ packages/sdk/src/nft/mint.ts
@@ -1,6 +1,7 @@
 import type { Fcl } from "@rarible/fcl-types"
 import type { Maybe } from "@rarible/types/build/maybe"
 import type { FlowRoyalty } from "@rarible/flow-api-client"
+import { toFlowAddress } from "@rarible/types"
 import type { AuthWithPrivateKey, FlowNetwork, FlowTransaction } from "../types"
 import { runTransaction, waitForSeal } from "../common/transaction"
 import { getNftCode } from "../tx-code-store/nft"
@@ -25,16 +26,21 @@ export async function mint(
 	if (!fcl) {
 		throw new Error("Fcl is required for mint")
 	}
-	const { map, address, features, name } = getCollectionConfig(
+	const from = auth ? toFlowAddress((await auth()).addr) : toFlowAddress((await fcl.currentUser().snapshot()).addr!)
+	if (!from) {
+		throw new Error("FLOW-SDK: Can't get current user address")
+	}
+	const { map, address, features, name, userCollectionId } = getCollectionConfig(
 		network,
 		collection,
 	)
+	const minterId = userCollectionId
 	if (features.includes("MINT")) {
 		const validatedRoyalties = validateRoyalties(royalties)
 		const txId = await runTransaction(
 			fcl,
 			map,
-			getNftCode(name).mint(fcl, address, metadata, validatedRoyalties),
+			getNftCode(name).mint({ fcl, address, metadata, royalties: validatedRoyalties, receiver: from, minterId }),
 			auth,
 		)
 		const txResult = await waitForSeal(fcl, txId)
@@ -50,5 +56,5 @@ export async function mint(
 		}
 		throw new Error("Something went wrong, transaction sent but events is empty")
 	}
-	throw new Error("This collection doesn't support 'mint'")
+	throw new Error("This collection doesn't support minting")
 }