+++ packages/sdk/src/order/fill/fill-bid-order.ts
@@ -1,15 +1,14 @@
 import type { Fcl } from "@rarible/fcl-types"
 import type { FlowAddress } from "@rarible/types"
-import type { AuthWithPrivateKey, FlowCurrency, FlowFee, FlowTransaction } from "../../types"
+import type { AuthWithPrivateKey, FlowCurrency, FlowFee, FlowTransaction, NonFungibleContract } from "../../types"
 import { runTransaction, waitForSeal } from "../../common/transaction"
-import type { FlowCollectionName } from "../../common/collection"
 import { getBidCode } from "../../tx-code-store/order/rarible-open-bid"
 
 export async function fillBidOrder(
 	fcl: Fcl,
 	auth: AuthWithPrivateKey,
 	currency: FlowCurrency,
-	name: FlowCollectionName,
+	name: NonFungibleContract,
 	map: Record<string, string>,
 	orderId: number,
 	owner: FlowAddress,