+++ packages/sdk/src/tx-code-store/order/prepare-order-code.test.ts
@@ -1,21 +1,20 @@
 import { openBidTransactionCode } from "@rarible/flow-sdk-scripts"
-import type { FlowCollectionName } from "../../common/collection"
-import type { FlowCurrency } from "../../types"
+import type { FlowCurrency, NonFungibleContract } from "../../types"
 import { prepareOrderCode } from "./prepare-order-code"
 
 describe("Get bid cadence code", () => {
 	test("should replace", () => {
-		const code = prepareOrderCode(openBidTransactionCode.openBid, "RaribleNFT" as FlowCollectionName, "FLOW")
+		const code = prepareOrderCode(openBidTransactionCode.openBid, "RaribleNFT", "FLOW")
 		expect(code.search("FlowToken")).toBeGreaterThan(0)
 		expect(code.search("flowToken")).toBeGreaterThan(0)
 		expect(code.search("RaribleNFT")).toBeGreaterThan(0)
 	})
 	test("should throw error wrong currency", () => {
-		const code = () => prepareOrderCode(openBidTransactionCode.openBid, "RaribleNFT" as FlowCollectionName, "ETH" as FlowCurrency)
+		const code = () => prepareOrderCode(openBidTransactionCode.openBid, "RaribleNFT", "ETH" as FlowCurrency)
 		expect(code).toThrow(Error)
 	})
 	test("should throw error wrong collection", () => {
-		const code = () => prepareOrderCode(openBidTransactionCode.openBid, "RaribleNF" as FlowCollectionName, "FLOW")
+		const code = () => prepareOrderCode(openBidTransactionCode.openBid, "RaribleNF" as NonFungibleContract, "FLOW")
 		expect(code).toThrow(Error)
 	})
 })