+++ packages/sdk/src/types.ts
@@ -12,17 +12,27 @@ export type FlowPayouts = FlowFee[]
 
 export type FungibleContracts = FUSDType | FlowTokenType
 
-export type NonFungibleContracts =
-	RaribleNFTType |
-	EvolutionType |
-	MotoGPCardType |
-	TopShotType |
-	MugenNFT |
-	CNN_NFT |
-	MatrixWorldFlowFestNFT |
-	MatrixWorldVoucher |
-	RaribleNFTv2 |
-	SoftCollection
+
+export const NON_FUNGIBLE_CONTRACTS = [
+	"RaribleNFT",
+	"Evolution",
+	"MotoGPCard",
+	"TopShot",
+	"MugenNFT",
+	"CNN_NFT",
+	"MatrixWorldFlowFestNFT",
+	"MatrixWorldVoucher",
+	"DisruptArt",
+	"Art",
+	"StarlyCard",
+	"OneFootballCollectible",
+	"ChainmonstersRewards",
+	"RaribleNFTv2",
+	"SoftCollection",
+] as const
+
+export type NonFungibleContracts = typeof NON_FUNGIBLE_CONTRACTS
+export type NonFungibleContract = NonFungibleContracts[number]
 
 export type FlowContractName =
 	| NonFungibleTokenType
@@ -34,7 +44,7 @@ export type FlowContractName =
 	| TopShotFeeType
 	| RaribleOpenBid
 	| FungibleContracts
-	| NonFungibleContracts
+	| NonFungibleContract
 	| MetadataViews
 
 export type AuthWithPrivateKey = undefined | ((account?: any) => Promise<any>)
@@ -55,14 +65,4 @@ export type RaribleOrderType = "RaribleOrder"
 export type LicensedNFTType = "LicensedNFT"
 export type TopShotFeeType = "TopShotFee"
 export type RaribleOpenBid = "RaribleOpenBid"
-export type MotoGPCardType = "MotoGPCard"
-export type EvolutionType = "Evolution"
-export type TopShotType = "TopShot"
-export type RaribleNFTType = "RaribleNFT"
-export type MugenNFT = "MugenNFT"
-export type CNN_NFT = "CNN_NFT"
-export type MatrixWorldFlowFestNFT = "MatrixWorldFlowFestNFT"
-export type MatrixWorldVoucher = "MatrixWorldVoucher"
-export type RaribleNFTv2 = "RaribleNFTv2"
-export type SoftCollection = "SoftCollection"
 export type MetadataViews = "MetadataViews"