+++ packages/test-common/src/config.ts
@@ -23,6 +23,12 @@ export const CONTRACTS: Record<Contracts, string> = {
 	CNN_NFT: "CNN_NFT",
 	MatrixWorldFlowFestNFT: "MatrixWorldFlowFestNFT",
 	MatrixWorldVoucher: "MatrixWorldVoucher",
+	DisruptArt: "DisruptArt",
+	Art: "Art",
+	StarlyCard: "StarlyCard",
+	OneFootballCollectible: "OneFootballCollectible",
+	ChainmonstersRewards: "ChainmonstersRewards",
+	Content: "Content",
 	SoftCollection: "SoftCollection",
 	RaribleNFTv2: "RaribleNFTv2",
 	MetadataViews: "MetadataViews",
@@ -49,6 +55,13 @@ type Contracts =
 	"MugenNFT" |
 	"CNN_NFT" |
 	"MatrixWorldFlowFestNFT" |
+	"MatrixWorldVoucher" |
+	"DisruptArt" |
+	"Art" |
+	"StarlyCard" |
+	"OneFootballCollectible" |
+	"ChainmonstersRewards" |
+	"Content" |
 	"MatrixWorldVoucher" |
 	"SoftCollection" |
 	"RaribleNFTv2" |